Changed infrastructure renamed files.
