+++ robupy/minimize_scalar.py
